% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F3_Correlation.R
\name{bBallCorr}
\alias{bBallCorr}
\title{Function 3: returns a correlation matrix of numeric player statistics in a given year}
\usage{
bBallCorr(x)
}
\description{
This function allows you to find Jeremy Lin's stats for a given year.
}
\examples{
bBallCorr(2014)
}
