% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1_JeremyLin.R
\name{jeremyLin}
\alias{jeremyLin}
\title{Function 1: Jeremy Lin stats}
\usage{
jeremyLin(x)
}
\description{
This function allows you to find Jeremy Lin's stats for a given year.
}
\examples{
jeremyLin(2014)
}
