% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F2_oldestPlayer.R
\name{oldestPlayer}
\alias{oldestPlayer}
\title{Function 2: returns oldest player in the NBA in a given year}
\usage{
oldestPlayer(x)
}
\description{
This function allows you to find Jeremy Lin's stats for a given year.
}
\examples{
oldestPlayer(2014)
}
